fixed migration and issues #8, #9

+++ db/migrate/20160224233838_delete_unnecessary_columns_from_membros.rb
@@ -1,8 +1,8 @@
 class DeleteUnnecessaryColumnsFromMembros < ActiveRecord::Migration
   def change
-    remove_column :membros, :titulo_eleitor_numero_inscricao, :string
-    remove_column :membros, :titulo_eleitor_zona, :string
-    remove_column :membros, :titulo_eleitor_secao, :string
-    remove_column :membros, :titulo_eleitor_data_emissao, :date
+    remove_column :membros, :titulo_eleitor_numero_inscricao, :string if Membro.column_names.include?('titulo_eleitor_numero_inscricao')
+    remove_column :membros, :titulo_eleitor_zona, :string if Membro.column_names.include?('titulo_eleitor_zona')
+    remove_column :membros, :titulo_eleitor_secao, :string if Membro.column_names.include?('titulo_eleitor_secao')
+    remove_column :membros, :titulo_eleitor_data_emissao, :date if Membro.column_names.include?('titulo_eleitor_data_emissao')
   end
 end
